Migrate st_lawrence (rotated-SOVC PDF, extract_text row parse)
